added function pointers for each OpenSHMEM routine we are using. some of these are likely incorrect. these pointers will be overwitten with the actual openshmem routine if it exists. if not, then it will remain a null pointer and that routine won't be tested. this way we don't get compiler errors if a routine is not implemented. now i need to change all the tests so they use the function pointers instead of the routines directly

+++ src/tests/atomics/atomics_tests.cpp
@@ -10,10 +10,10 @@
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_fetch() {
-  static int *dest;
-  static int fetch;
-  dest = (int *)shmem_malloc(sizeof(int));
-  int value = 42;
+  static long *dest;
+  static long fetch;
+  dest = (long *)shmem_malloc(sizeof(long));
+  long value = 42;
   *dest = value;
   shmem_barrier_all();
   fetch = shmem_atomic_fetch(dest, 1);
@@ -27,9 +27,9 @@ bool test_shmem_atomic_fetch() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_set() {
-  static int *dest;
-  dest = (int *)shmem_malloc(sizeof(int));
-  int value = 42;
+  static long *dest;
+  dest = (long *)shmem_malloc(sizeof(long));
+  long value = 42;
   shmem_barrier_all();
   shmem_atomic_set(dest, value, 1);
   shmem_barrier_all();
@@ -43,12 +43,12 @@ bool test_shmem_atomic_set() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_compare_swap() {
-  static int *dest;
-  dest = (int *)shmem_malloc(sizeof(int));
-  int old = 42, new_val = 43;
+  static long *dest;
+  dest = (long *)shmem_malloc(sizeof(long));
+  long old = 42, new_val = 43;
   *dest = old;
   shmem_barrier_all();
-  int swapped = shmem_atomic_compare_swap(dest, old, new_val, 1);
+  long swapped = shmem_atomic_compare_swap(dest, old, new_val, 1);
   bool success = (swapped == old && *dest == new_val);
   shmem_free(dest);
   return success;
@@ -59,12 +59,12 @@ bool test_shmem_atomic_compare_swap() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_swap() {
-  static int *dest;
-  dest = (int *)shmem_malloc(sizeof(int));
-  int value = 42, new_val = 43;
+  static long *dest;
+  dest = (long *)shmem_malloc(sizeof(long));
+  long value = 42, new_val = 43;
   *dest = value;
   shmem_barrier_all();
-  int swapped = shmem_atomic_swap(dest, new_val, 1);
+  long swapped = shmem_atomic_swap(dest, new_val, 1);
   bool success = (swapped == value && *dest == new_val);
   shmem_free(dest);
   return success;
@@ -75,10 +75,10 @@ bool test_shmem_atomic_swap() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_fetch_inc() {
-  static int *dest;
-  static int fetch;
-  dest = (int *)shmem_malloc(sizeof(int));
-  int value = 42;
+  static long *dest;
+  static long fetch;
+  dest = (long *)shmem_malloc(sizeof(long));
+  long value = 42;
   *dest = value;
   shmem_barrier_all();
   fetch = shmem_atomic_fetch_inc(dest, 1);
@@ -92,9 +92,9 @@ bool test_shmem_atomic_fetch_inc() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_inc() {
-  static int *dest;
-  dest = (int *)shmem_malloc(sizeof(int));
-  int value = 42;
+  static long *dest;
+  dest = (long *)shmem_malloc(sizeof(long));
+  long value = 42;
   *dest = value;
   shmem_barrier_all();
   shmem_atomic_inc(dest, 1);
@@ -109,10 +109,10 @@ bool test_shmem_atomic_inc() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_fetch_add() {
-  static int *dest;
-  static int fetch;
-  dest = (int *)shmem_malloc(sizeof(int));
-  int value = 42, add_val = 10;
+  static long *dest;
+  static long fetch;
+  dest = (long *)shmem_malloc(sizeof(long));
+  long value = 42, add_val = 10;
   *dest = value;
   shmem_barrier_all();
   fetch = shmem_atomic_fetch_add(dest, add_val, 1);
@@ -126,9 +126,9 @@ bool test_shmem_atomic_fetch_add() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_add() {
-  static int *dest;
-  dest = (int *)shmem_malloc(sizeof(int));
-  int value = 42, add_val = 10;
+  static long *dest;
+  dest = (long *)shmem_malloc(sizeof(long));
+  long value = 42, add_val = 10;
   *dest = value;
   shmem_barrier_all();
   shmem_atomic_add(dest, add_val, 1);
@@ -143,10 +143,10 @@ bool test_shmem_atomic_add() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_fetch_and() {
-  static int *dest;
-  static int fetch;
-  dest = (int *)shmem_malloc(sizeof(int));
-  int value = 42, and_val = 15;
+  static long *dest;
+  static long fetch;
+  dest = (long *)shmem_malloc(sizeof(long));
+  long value = 42, and_val = 15;
   *dest = value;
   shmem_barrier_all();
   fetch = shmem_atomic_fetch_and(dest, and_val, 1);
@@ -160,9 +160,9 @@ bool test_shmem_atomic_fetch_and() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_and() {
-  static int *dest;
-  dest = (int *)shmem_malloc(sizeof(int));
-  int value = 42, and_val = 15;
+  static long *dest;
+  dest = (long *)shmem_malloc(sizeof(long));
+  long value = 42, and_val = 15;
   *dest = value;
   shmem_barrier_all();
   shmem_atomic_and(dest, and_val, 1);
@@ -177,10 +177,10 @@ bool test_shmem_atomic_and() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_fetch_or() {
-  static int *dest;
-  static int fetch;
-  dest = (int *)shmem_malloc(sizeof(int));
-  int value = 42, or_val = 15;
+  static long *dest;
+  static long fetch;
+  dest = (long *)shmem_malloc(sizeof(long));
+  long value = 42, or_val = 15;
   *dest = value;
   shmem_barrier_all();
   fetch = shmem_atomic_fetch_or(dest, or_val, 1);
@@ -194,9 +194,9 @@ bool test_shmem_atomic_fetch_or() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_or() {
-  static int *dest;
-  dest = (int *)shmem_malloc(sizeof(int));
-  int value = 42, or_val = 15;
+  static long *dest;
+  dest = (long *)shmem_malloc(sizeof(long));
+  long value = 42, or_val = 15;
   *dest = value;
   shmem_barrier_all();
   shmem_atomic_or(dest, or_val, 1);
@@ -211,10 +211,10 @@ bool test_shmem_atomic_or() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_fetch_xor() {
-  static int *dest;
-  static int fetch;
-  dest = (int *)shmem_malloc(sizeof(int));
-  int value = 42, xor_val = 15;
+  static long *dest;
+  static long fetch;
+  dest = (long *)shmem_malloc(sizeof(long));
+  long value = 42, xor_val = 15;
   *dest = value;
   shmem_barrier_all();
   fetch = shmem_atomic_fetch_xor(dest, xor_val, 1);
@@ -228,9 +228,9 @@ bool test_shmem_atomic_fetch_xor() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_xor() {
-  static int *dest;
-  dest = (int *)shmem_malloc(sizeof(int));
-  int value = 42, xor_val = 15;
+  static long *dest;
+  dest = (long *)shmem_malloc(sizeof(long));
+  long value = 42, xor_val = 15;
   *dest = value;
   shmem_barrier_all();
   shmem_atomic_xor(dest, xor_val, 1);
@@ -245,10 +245,10 @@ bool test_shmem_atomic_xor() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_fetch_nbi() {
-  static int *dest;
-  static int fetch;
-  dest = (int *)shmem_malloc(sizeof(int));
-  int value = 42;
+  static long *dest;
+  static long fetch;
+  dest = (long *)shmem_malloc(sizeof(long));
+  long value = 42;
   *dest = value;
   shmem_barrier_all();
   shmem_atomic_fetch_nbi(&fetch, dest, 1);
@@ -263,11 +263,11 @@ bool test_shmem_atomic_fetch_nbi() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_compare_swap_nbi() {
-  static int *dest;
-  static int fetch;
-  dest = (int *)shmem_malloc(sizeof(int));
+  static long *dest;
+  static long fetch;
+  dest = (long *)shmem_malloc(sizeof(long));
   fetch = 0;
-  int old = 42, new_val = 43;
+  long old = 42, new_val = 43;
   *dest = old;
   shmem_barrier_all();
   shmem_atomic_compare_swap_nbi(&fetch, dest, old, new_val, 1);
@@ -282,11 +282,11 @@ bool test_shmem_atomic_compare_swap_nbi() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_swap_nbi() {
-  static int *dest;
-  static int fetch;
-  dest = (int *)shmem_malloc(sizeof(int));
+  static long *dest;
+  static long fetch;
+  dest = (long *)shmem_malloc(sizeof(long));
   fetch = 0;
-  int value = 42, new_val = 43;
+  long value = 42, new_val = 43;
   *dest = value;
   shmem_barrier_all();
   shmem_atomic_swap_nbi(&fetch, dest, new_val, 1);
@@ -301,11 +301,11 @@ bool test_shmem_atomic_swap_nbi() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_fetch_inc_nbi() {
-  static int *dest;
-  static int fetch;
-  dest = (int *)shmem_malloc(sizeof(int));
+  static long *dest;
+  static long fetch;
+  dest = (long *)shmem_malloc(sizeof(long));
   fetch = 0;
-  int value = 42;
+  long value = 42;
   *dest = value;
   shmem_barrier_all();
   shmem_atomic_fetch_inc_nbi(&fetch, dest, 1);
@@ -320,11 +320,11 @@ bool test_shmem_atomic_fetch_inc_nbi() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_fetch_add_nbi() {
-  static int *dest;
-  static int fetch;
-  dest = (int *)shmem_malloc(sizeof(int));
+  static long *dest;
+  static long fetch;
+  dest = (long *)shmem_malloc(sizeof(long));
   fetch = 0;
-  int value = 42, add_val = 10;
+  long value = 42, add_val = 10;
   *dest = value;
   shmem_barrier_all();
   shmem_atomic_fetch_add_nbi(&fetch, dest, add_val, 1);
@@ -339,11 +339,11 @@ bool test_shmem_atomic_fetch_add_nbi() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_fetch_and_nbi() {
-  static int *dest;
-  static int fetch;
-  dest = (int *)shmem_malloc(sizeof(int));
+  static long *dest;
+  static long fetch;
+  dest = (long *)shmem_malloc(sizeof(long));
   fetch = 0;
-  int value = 42, and_val = 15;
+  long value = 42, and_val = 15;
   *dest = value;
   shmem_barrier_all();
   shmem_atomic_fetch_and_nbi(&fetch, dest, and_val, 1);
@@ -358,11 +358,11 @@ bool test_shmem_atomic_fetch_and_nbi() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_fetch_or_nbi() {
-  static int *dest;
-  static int fetch;
-  dest = (int *)shmem_malloc(sizeof(int));
+  static long *dest;
+  static long fetch;
+  dest = (long *)shmem_malloc(sizeof(long));
   fetch = 0;
-  int value = 42, or_val = 15;
+  long value = 42, or_val = 15;
   *dest = value;
   shmem_barrier_all();
   shmem_atomic_fetch_or_nbi(&fetch, dest, or_val, 1);
@@ -377,11 +377,11 @@ bool test_shmem_atomic_fetch_or_nbi() {
   @return True if the test is successful, false otherwise.
  */
 bool test_shmem_atomic_fetch_xor_nbi() {
-  static int *dest;
-  static int fetch;
-  dest = (int *)shmem_malloc(sizeof(int));
+  static long *dest;
+  static long fetch;
+  dest = (long *)shmem_malloc(sizeof(long));
   fetch = 0;
-  int value = 42, xor_val = 15;
+  long value = 42, xor_val = 15;
   *dest = value;
   shmem_barrier_all();
   shmem_atomic_fetch_xor_nbi(&fetch, dest, xor_val, 1);
